backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
qdrant-client>=1.9.0
cohere>=4.4.0
openai>=1.10.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
python-multipart>=0.0.6